LayoutTests/http/tests/security/javascriptURL/xss-DENIED-from-javascript-url-in-foreign-domain-window-open-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-from-javascript-url-in-foreign-domain-window-open.html from frame with URL about:blank. Domains, protocols and ports must match.

Opener Frame

Pass: Cross frame access from a javascript: URL on a different domain was denied.
